Time Contextual Headers [1/2]
This commit adds image headers to the notification panel header,
based on the date/time.

There is a default provider that provides sky pictures
as well as special events like Christmas and new
year's eve.

More will be coming, with eventually a way to use an external
(as in, Internet) provider for those pictures.

Change-Id: Ic26c0f9fef663213f942f864f7f2dd4b14dfaec1

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/StatusBarHeaderView.java