OMS7-N: Ensure unique type IDs when parsing overlays [6/14]
Always add a block to the idmap for each type in the overlay package
even if there are no matching resources for that type. This ensures that
ResTable::parsePackage will always get a unique ID for each type.

An alternative implementation would be to skip non-matching types when
parsing the overlay package, but that would prevent resolving references
to resources in XML.

Change-Id: Ia405fa53562967139965c4c6aa49f252b79d14d2
Co-authored-by: Anton Risberg Alakula <anton.risbergalakula@sonymobile.com>