QS: Smart Ambient-LiftToWake tile [1/2]
Useful when cleaning up the device screen :D

Now that N makes available two separate option for doze notifications
(display-ambient and Gestures-LiftToWake) we need to handle them both
with the tile.
It will act this way:
1) if any of the two options is enabled, and you tap the tile,
both options are disabled so you can clean the screen or avoid
unwanted disturbing wakes;
2) if you tap the tile again, the tile will restore, separately,
the two options state you had before disabling them;
3) if you manually disable the two options, the tile will go to
 disabled state;
4) if both options (thus the tile) are disabled, and you manually enable
on of them, the tile will go to enabled state and act again as 1) and 2)

Thanks Alex Cruz (Mazda-- ) for the icons!

Change-Id: I344a254faec9ba6ae47d3afceeb84f8175fec097