base: Introduce Accidental Touch
* Citrus-CAF:
  Forward-port and adapt to Android Pie

Accidental Touch is a feature useful for many, especially the gamers out there.
It prevents any sort of accidental touch on hardware buttons while the touchscreen is
in use.
This means you won’t experience any unexpected behaviour while playing games,
web browsing or even when attempting to reach your finger out to the edge of
that humongous display.

Change-Id: I7ec9f001f76b65d6cec74caa64e8b3890780ee94
Signed-off-by: Adarsh-MR <adarshmr1998@gmail.com>