SystemUI: Status bar clock customization [1/2]
Adds the following customization options to the status bar clock:
- Clock position
- Show seconds
- Show AM/PM
- Show day of the week

Removed the unused AOSP code that allowed changing of the AM/PM
size. Instead, both AM/PM and day of the week is set to be 0.75x the
default. Now that it is ≈ the average of the original "small" and
"normal" options, make it boolean enabled along with the others.

This is a modified combination of commits by the following people:
- Lars Greiss  <kufikugel@googlemail.com>
- mydongistiny <jaysonedson@gmail.com>
- Joe Maples   <joe@frap129.org>

Test: Look at the status bar when testing each option. Also make
      sure that the items are hidden properly when heads up is
      showing, then make sure they reappear when heads up is hiding.

Change-Id: Iaa180159ad93c4aa8b8032c7a1258ab3f341c5dd
Signed-off-by: Adarsh-MR <adarshmr1998@gmail.com>